% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geography.R
\name{add_geocodes}
\alias{add_geocodes}
\title{Add geographical area codes to a dataset, starting from a postcode column}
\usage{
add_geocodes(
  dt,
  clean_pc = TRUE,
  pc_cname = "postcode",
  oa_only = FALSE,
  add_oa = TRUE,
  census = TRUE,
  admin = TRUE,
  postal = TRUE,
  electoral = FALSE,
  nhs = FALSE,
  crime = FALSE,
  cols_in = NULL,
  cols_out = NULL
)
}
\arguments{
\item{dt}{a data.table}

\item{clean_pc}{if the postcode column needs to be cleaned beforehand}

\item{pc_cname}{if clean_pc is TRUE, this is the column name to consider}

\item{oa_only}{Add only the Output Area column, disregarding all the other options}

\item{census}{Add geographies related to the "Census" hierarchy: 'LSOA', 'MSOA', 'LAD'}

\item{admin}{Add geographies related to the "Admin" hierarchy: 'LAD', 'CTY', 'RGN', 'CTRY'}

\item{postal}{Add geographies related to the "Postal" hierarchy: 'PCS', 'PCD', 'PCT', 'PCA'}

\item{electoral}{Add geographies related to the "Electoral" hierarchy: 'PCON', 'WARD', 'CED'}

\item{nhs}{Add geographies related to the "NHS" hierarchy: 'CCG', 'NHSO', 'NHSR'}

\item{crime}{Add geographies related to the "Police" hierarchy: 'CSP', 'PFA'}

\item{cols_in}{Insert here isolated columns to add to the output dataset}

\item{cols_out}{The columns you don't want to be included in the output Note that you can not exclude neither OA nor WPZ.}
}
\value{
a data.table with possibly cleaned postcode, OA and WPZ columns, plus all other specified
}
\description{
Add geographical area codes to a dataset, starting from a postcode column
}
\author{
Luca Valnegri, \email{l.valnegri@datamaps.co.uk}
}
