% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{add_Kcomma}
\alias{add_Kcomma}
\title{Add a comma to thousands}
\usage{
add_Kcomma(x)
}
\arguments{
\item{x}{a numeric vector}
}
\value{
A character vector
}
\description{
Add a comma to thousands
}
\author{
Luca Valnegri, \email{l.valnegri@datamaps.co.uk}
}
