% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps_leaflet.R
\name{add_maptile}
\alias{add_maptile}
\title{add_maptile}
\usage{
add_maptile(m, x, grp = NULL)
}
\arguments{
\item{m}{a `leaflet` object}

\item{x}{text or url description of the required maptile (see \code{tiles.lst})}

\item{grp}{an optional group name for the layer}
}
\value{
A `leaflet` object
}
\description{
Add a background layer to a \code{leaflet} map
}
\author{
Luca Valnegri, \email{l.valnegri@datamaps.co.uk}
}
