% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postcodes.R
\name{add_postcodes_coords}
\alias{add_postcodes_coords}
\title{Add to a dataset with a postcodes column the corresponding geographic coordinates}
\usage{
add_postcodes_coords(x, cname = "PCU")
}
\arguments{
\item{x}{a data.table}

\item{cname}{the column containing the postcodes}
}
\value{
a data.table
}
\description{
Add to a dataset with a postcodes column the corresponding geographic coordinates
}
\author{
Luca Valnegri, \email{l.valnegri@datamaps.co.uk}
}
