% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{add_space}
\alias{add_space}
\title{Add a trailing space to a value which is less than 10}
\usage{
add_space(x)
}
\arguments{
\item{x}{a numeric value}
}
\value{
A character value
}
\description{
Add a trailing space to a value which is less than 10
}
\author{
Luca Valnegri, \email{l.valnegri@datamaps.co.uk}
}
