% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{add_tile}
\alias{add_tile}
\title{Add a background layer to a \code{leaflet} map}
\usage{
add_tile(m, x)
}
\arguments{
\item{m}{a \code{leaflet} object}

\item{x}{text or url description of the required maptile (see \code{tiles.lst})}
}
\value{
A \code{leaflet} object
}
\description{
Add a background layer to a \code{leaflet} map
}
\author{
Luca Valnegri, \email{l.valnegri@datamaps.co.uk}
}
