% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{app_path}
\alias{app_path}
\title{Root of source data for any Shiny app}
\format{An object of class \code{character} of length 1.}
\usage{
app_path
}
\description{
Root of source data for any Shiny app
}
\keyword{datasets}
