% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{basemap}
\alias{basemap}
\title{A leaflet map to start with, centered around a givcen coordinate or fitting a given bounding box}
\usage{
basemap(coords = c(-2.903205, 54.17463), bbox = NULL, zm = 6,
  tiles = tiles.lst)
}
\arguments{
\item{coords}{Center the map around this long/lat point (default is UK centroid)}

\item{bbox}{Alternative to <coords>, set the bounds of a map. In a 2x2 matrix form, long as first row, lat as second row; min as first  column, max as second column}

\item{zm}{The admissible minimum zoom. Default is 6, which is the minimum for the UK to be visible as a whole}

\item{tiles}{List of map backgroud tiles, by default all layers listed in tiles.lst. Pass a NULL value to add only the default tile.}
}
\value{
a leaflet map
}
\description{
A leaflet map to start with, centered around a givcen coordinate or fitting a given bounding box
}
\examples{
\dontrun{
    example code
}

}
\references{
leaflet, maps
}
\author{
Luca Valnegri, \email{l.valnegri@datamaps.co.uk}
}
