% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{bbox.it}
\alias{bbox.it}
\title{Italian Bounding Box}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 2 rows and 2 columns.
}
\usage{
bbox.it
}
\description{
Italian Bounding Box
}
\keyword{datasets}
