% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{bbox.uk.cut}
\alias{bbox.uk.cut}
\title{UK Bounding Box [No Outer Islands]}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 2 rows and 2 columns.
}
\usage{
bbox.uk.cut
}
\description{
The Bounding Box for the UK without the four following Islands:
 - S12000013 Na h-Eileanan Siar
 - S12000023 Orkney Islands
 - S12000027 Shetland Islands
 - E06000053 Isles of Scilly
}
\details{
longitude: -8.178225 to  1.76368
latitude:  49.958709 to 58.69690
}
\keyword{datasets}
