% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{bbox.uk}
\alias{bbox.uk}
\title{UK Bounding Box}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 2 rows and 2 columns.
}
\usage{
bbox.uk
}
\description{
longitude: -8.650007 to  1.76368
latitude:  49.882346 to 60.84567
}
\keyword{datasets}
