% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geography.R
\name{bnd_merge}
\alias{bnd_merge}
\title{Merge multiple area polygons into one single SpatialPolygonDataFrame}
\usage{
bnd_merge(ids, cpath)
}
\arguments{
\item{ids}{}
}
\value{
a SpatialPolygonDataFrame
}
\description{
Merge multiple area polygons into one single SpatialPolygonDataFrame
}
\examples{
\dontrun{
}

}
\author{
Luca Valnegri, \email{l.valnegri@datamaps.co.uk}
}
