% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geography.R
\name{bnd_merge_path}
\alias{bnd_merge_path}
\title{Append together multiple SpatialPolygonDataFrame into one single SpatialPolygonDataFrame}
\usage{
bnd_merge_path(fnames, bpath)
}
\arguments{
\item{fnames}{the list of names (or ids) of the boundaries to be "merged" together}

\item{bpath}{the directory where ALL the boundaries are stored}
}
\value{
a SpatialPolygonDataFrame
}
\description{
Append together multiple SpatialPolygonDataFrame into one single SpatialPolygonDataFrame
}
\author{
Luca Valnegri, \email{l.valnegri@datamaps.co.uk}
}
