% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{bnd_path}
\alias{bnd_path}
\title{Location of the top boundaries folder in rds format in the shared repository}
\format{
An object of class \code{character} of length 1.
}
\usage{
bnd_path
}
\description{
Location of the top boundaries folder in rds format in the shared repository
}
\keyword{datasets}
