% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{bndit_path}
\alias{bndit_path}
\title{Location of the Italian boundaries in the shared repository}
\format{
An object of class \code{character} of length 1.
}
\usage{
bndit_path
}
\description{
Location of the Italian boundaries in the shared repository
}
\keyword{datasets}
