% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{bndit_spath}
\alias{bndit_spath}
\title{Location of the Italian boundaries, in the shared repository, typically used in R applications (= rds format)}
\format{
An object of class \code{character} of length 1.
}
\usage{
bndit_spath
}
\description{
Location of the Italian boundaries, in the shared repository, typically used in R applications (= rds format)
}
\keyword{datasets}
