% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{bnduk_path}
\alias{bnduk_path}
\title{Location of the boundaries for the UK in the shared repository}
\format{An object of class \code{character} of length 1.}
\usage{
bnduk_path
}
\description{
Location of the boundaries for the UK in the shared repository
}
\keyword{datasets}
