% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geography.R
\name{build_lookups_table}
\alias{build_lookups_table}
\title{Build a lookup table child <=> parent using the postcodes table from the ONS geography database
This function should not be used with 'OA' as child because in the csv files from ONS there are 265 OAs missing (36 ENG, 229 SCO)
Always remember to check column 'pct_coverage' for values less than 100}
\usage{
build_lookups_table(child, parent, is_active = TRUE,
  filter_country = NULL, save_results = FALSE,
  out_path = file.path(ext_path, "geography", "lookups"))
}
\arguments{
\item{child}{the code for the lower level geography}

\item{parent}{the code for the higher level geography}

\item{is_active}{if TRUE, keep only live postcodes for the calculation}

\item{filter_country}{indicates if the calculation must be done on less than the UK}

\item{save_results}{if TRUE, the result dataset will also be saved}

\item{out_path}{if save_results is TRUE, the folder where to save the output file (which will be called "paste0(child, '_to_', parent))"}
}
\value{
a data.table with two columns
}
\description{
Build a lookup table child <=> parent using the postcodes table from the ONS geography database
This function should not be used with 'OA' as child because in the csv files from ONS there are 265 OAs missing (36 ENG, 229 SCO)
Always remember to check column 'pct_coverage' for values less than 100
}
\examples{
\dontrun{
  build_lookups_table('LSOA', 'MSOA')
  build_lookups_table('LSOA', 'CTY', filter_country = 'E')
}

}
\author{
Luca Valnegri, \email{l.valnegri@datamaps.co.uk}
}
