% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{capitalize}
\alias{capitalize}
\title{Capitalize the first only or all the words in a text column, meanwhile cleaning of redundant spaces}
\usage{
capitalize(dt, x, all_words = TRUE, as_factor = TRUE)
}
\arguments{
\item{dt}{a data.table}

\item{x}{the column whose valued have to be cleaned and capitalized}

\item{all_words}{if TRUE the capitalization is applied on all the words of the content. When FALSE only the first word is capitalized.}

\item{as_factor}{if TRUE, the specified column is converted to factor}
}
\value{
A data.table similar to the one in input, but with the specified column modified as asked
}
\description{
Capitalize the first only or all the words in a text column, meanwhile cleaning of redundant spaces
}
\author{
Luca Valnegri, \email{l.valnegri@datamaps.co.uk}
}
