% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{center.it}
\alias{center.it}
\title{Italian Centroid (Narni, Terni, Umbria)}
\format{
An object of class \code{numeric} of length 2.
}
\usage{
center.it
}
\description{
Italian Centroid (Narni, Terni, Umbria)
}
\keyword{datasets}
