% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geography.R
\name{check_area_ids}
\alias{check_area_ids}
\title{Check if one or more strings are valid location ids, returning the correct common area type}
\usage{
check_area_ids(ids, give_warning = FALSE)
}
\arguments{
\item{ids}{a vector containing location ids of one unspecified location type}

\item{give_warning}{if there should be any explanation for the outcome}
}
\value{
a list with five components:
the valid ids, the invalid ids, the area type, the suffix name for the indexed location files, the parent id
}
\description{
Check if one or more strings are valid location ids, returning the correct common area type
}
\examples{
\dontrun{
}

}
\author{
Luca Valnegri, \email{l.valnegri@datamaps.co.uk}
}
