% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{class.lst}
\alias{class.lst}
\title{List of available methods when classifying numeric variables}
\format{
An object of class \code{character} of length 7.
}
\usage{
class.lst
}
\description{
List of available methods when classifying numeric variables
}
\keyword{datasets}
