% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{class.methods}
\alias{class.methods}
\title{List of available methods when classifying numeric variables}
\format{An object of class \code{character} of length 7.}
\usage{
class.methods
}
\description{
List prepared to be used as choices in Shiny input controls
}
\keyword{datasets}
