% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postcodes.R
\name{clean_postcode_dt}
\alias{clean_postcode_dt}
\title{Check a character string is a UK postcode, then convert it to a 7-char format}
\usage{
clean_postcode_dt(dt, cname = "postcode")
}
\arguments{
\item{dt}{a data.table}

\item{cname}{the column of <dt> to be checked and converted}
}
\value{
A data.table
}
\description{
The UK postcode system is hierarchical, the top level being "Postcode Area" (PCA) identified by 1 or 2 alphabetical character.
The next level is the "Postcode District" (PCD), also commonly known as the "outcode", and can take on several different formats, and anywhere from 2 to 4 alphanumeric characters long.
Next comes the Postcode Sector" (PCS), always identified by a single number, then finally the "unit", always formed by two alphabetical characters.
The combination of "sector" and "unit" is often called "incode", which is always 1 numeric character followed by 2 alphabetical characters.
}
\details{
Faster version than <clean_postcode>, returns a data.table instead of directly updating its argument
}
\author{
Luca Valnegri, \email{LucaValnegri@theambassadors.com}
}
