% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IT.R
\name{cmn2prv}
\alias{cmn2prv}
\title{cmn2prv}
\usage{
cmn2prv(x, as_char = TRUE)
}
\arguments{
\item{x}{The District code}

\item{as_char}{Indicate if the Province code should be returned as a 3-char code (default) or as an integer}
}
\value{
A string or an integer depending on the value of the parameter \code{as_char}
}
\description{
Given the ISTAT code for an Italian District, build the ISTAT code for the correspondent Province
}
\author{
Luca Valnegri, \email{l.valnegri@datamaps.co.uk}
}
