% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geography.R
\name{conv2spdf}
\alias{conv2spdf}
\title{Convert a dataframe in SpatialPointsDataFrame}
\usage{
conv2spdf(
  x,
  cid = "id",
  clon = "x_lon",
  clat = "y_lat",
  output = "spdf",
  ENtoLL = FALSE,
  crs = crs.gb
)
}
\arguments{
\item{x}{a dataframe with at least 3 columns, one for the id, the other two for the coordinates}

\item{cid}{the name of the id column}

\item{clon}{the name of the longitude or Easting column}

\item{clat}{the name of the latitude or Northing column}

\item{output}{Specify which output to return in case ENtoLL is TRUE: spdf, df, mgdf, map.}

\item{ENtoLL}{When TRUE, it first apply a EN transformation, before calculating geographical coordinates}

\item{crs}{if ENtoLL is true, the CRS to apply to x to calculate longitude and latitude}
}
\value{
a SpatialPointsDataFrame in WGS84, a dataframe, a leaflet map object
}
\description{
Convert a dataframe in SpatialPointsDataFrame
}
\author{
Luca Valnegri, \email{l.valnegri@datamaps.co.uk}
}
