% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangling.R
\name{convert_vars}
\alias{convert_vars}
\title{Convert a set of variables in the specified datatype}
\usage{
convert_vars(dt, vset, vtype)
}
\arguments{
\item{dt}{a data.table}

\item{vset}{the variables to convert}

\item{vtype}{the datatype to convert to}
}
\value{
None (albeit the data.table in input is modified with the content of the specified column cleaned and converted)
}
\description{
Convert a set of variables in the specified datatype
}
\examples{
\dontrun{
  convert_vars(dts, c('type', 'phase'), 'integer')
}

}
\author{
Luca Valnegri, \email{l.valnegri@datamaps.co.uk}
}
