% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{copy_app_code}
\alias{copy_app_code}
\title{copy_app_code}
\usage{
copy_app_code(name, dest = "/srv/shiny-server/")
}
\arguments{
\item{name}{Name to assign to the subfolder in the Server Shiny (not necessarily equal to package name)}

\item{dest}{Shiny Server root folder}
}
\value{
None
}
\description{
Copy the content of the \code{shinyapp} folder inside a package into a Shiny Server subfolder named \code{name}.
}
\author{
Luca Valnegri, \email{l.valnegri@datamaps.co.uk}
}
