% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangling.R
\name{create_age_class}
\alias{create_age_class}
\title{Create a new class variable, usually using a single year age variable}
\usage{
create_age_class(dt, cname = "age", newcol = NA, bin = 5)
}
\arguments{
\item{dt}{a data.table}

\item{cname}{the name of the column to be scanned to create the class}

\item{newcol}{the name of the new categorical column}

\item{bin}{the width of each class, apart from the highest one whose width depends on the actual values of `cname`}
}
\value{
a data.table
}
\description{
Create a new class variable, usually using a single year age variable
}
\author{
Luca Valnegri, \email{l.valnegri@datamaps.co.uk}
}
