% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databases.R
\name{create_db}
\alias{create_db}
\title{Creates a MySQL database}
\usage{
create_db(x)
}
\arguments{
\item{x}{name of the database to be created}
}
\value{
None
}
\description{
\code{create_db} creates a database in the local MySQL Server
}
\examples{
\dontrun{
  create_db('helpers')
}

}
\author{
Luca Valnegri, \email{l.valnegri@datamaps.co.uk}
}
