% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databases.R
\name{create_dbtable}
\alias{create_dbtable}
\title{Creates a MySQL table}
\usage{
create_dbtable(tname, dname, tdef, dts = NULL)
}
\arguments{
\item{tname}{the name of the table to be created}

\item{dname}{the name of the database in which create the table}

\item{tdef}{the definitions for all the columns in the table}

\item{dts}{the optional dataset to be uploaded}
}
\value{
None
}
\description{
\code{create_dbtable} creates a table in a specified database in the local MySQL Server
}
\examples{

\dontrun{
  x <- "
    name CHAR(25) NOT NULL,
    family CHAR(25) NOT NULL,
    package CHAR(25) NOT NULL,
    is_bold TINYINT(1) UNSIGNED NOT NULL,
    is_italic TINYINT(1) UNSIGNED NOT NULL,
    is_active TINYINT(1) UNSIGNED NOT NULL
  "
  create_dbtable('fonts', 'helpers', x)
}

\dontrun{
  x <- "
    ordering TINYINT(2) NOT NULL,
    provider CHAR(15) NOT NULL,
    name CHAR(20) NOT NULL,
    url CHAR(150) NOT NULL,
    attribution CHAR(250) NOT NULL,
    require_reg TINYINT(1) NOT NULL,
    max_zoom TINYINT(2) NOT NULL
  "
  y <- fread(file.path(pub_path, 'ancillaries', 'helpers', 'maptiles.csv'))
  create_dbtable('maptiles', 'helpers', x, y)
}

}
\author{
Luca Valnegri, \email{l.valnegri@datamaps.co.uk}
}
