% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_list}
\alias{create_list}
\title{create_list}
\usage{
create_list(x, nested = FALSE)
}
\arguments{
\item{x}{a data frame conveniently ordered, with two or three columns 
depending on whether the list should be simple or hierarchic}

\item{nested}{if `TRUE` the output is a hierarchical list, 
otherwise the list is a simple named one}
}
\value{
a list
}
\description{
Create a list of inputs, tipically to be used as a control for Shiny/RMarkdown
}
\author{
Luca Valnegri, \email{l.valnegri@datamaps.co.uk}
}
