% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geography.R
\name{crop_uk}
\alias{crop_uk}
\title{Clip (crop) a SpatialPolygonsDataframe against the UK extent (with or without Northern Scotland Islands)}
\usage{
crop_uk(bnd, crop_islands = TRUE)
}
\arguments{
\item{bnd}{the SpatialPolygonsDataframe to be clipped}

\item{crop_islands}{if the extent should exlude the northern Scotland isles}
}
\value{
a SpatialPolygonsDataframe
}
\description{
Clip (crop) a SpatialPolygonsDataframe against the UK extent (with or without Northern Scotland Islands)
}
\author{
Luca Valnegri, \email{l.valnegri@datamaps.co.uk}
}
