% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{crs.gb}
\alias{crs.gb}
\title{Coordinate Reference Systems (CRS) in EPSG code for OSGB 1936 / OSGB 1936 / Airy 1830 / British National Grid}
\format{An object of class \code{character} of length 1.}
\usage{
crs.gb
}
\description{
Coordinate system: Cartesian 2D CS
Axes: easting, northing (E,N)
proj4js: "+proj=tmerc +lat_0=49 +lon_0=-2 +k=0.9996012717 +x_0=400000 +y_0=-100000 +ellps=airy +datum=OSGB36 +units=m +no_defs";
Bounds: [WGS84] ; [Projected]
Centroid: [WGS84] ; [Projected]
Unit: metre
Area of use: Ireland - onshore.
Accuracy: 1m
}
\details{
sp::CRS(crs_code) to see the proj4js format for the specified epsg code
rgdal::make_EPSG() to build a table with code, name and proj4js format for all epsg codes
}
\keyword{datasets}
