% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{crs.it}
\alias{crs.it}
\title{Coordinate Reference Systems (CRS) in EPSG code (23032) for ED50 / UTM zone 32N - Denmark}
\format{
An object of class \code{character} of length 1.
}
\usage{
crs.it
}
\description{
Coordinate system: Cartesian 2D CS
Axes: easting, northing (E,N)
proj4js: "+proj=utm +zone=32 +ellps=intl +units=m +no_defs ";
Bounds: \link{WGS84} ; \link{Projected}
Centroid: \link{WGS84} ; \link{Projected}
Unit: metre
Area of use: Austria; Belgium; Denmark; Finland; Faroe islands; France; Germany (west); Gibraltar; Greece; Italy; Luxembourg; Netherlands; Norway; Portugal; Spain; Sweden; Switzerland.
Accuracy: various
}
\details{
sp::CRS(crs_code) to see the proj4js format for the specified epsg code
rgdal::make_EPSG() to build a table with code, name and proj4js format for all epsg codes
}
\keyword{datasets}
