% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{crs.ni}
\alias{crs.ni}
\title{Coordinate Reference Systems (CRS) in EPSG code for TM65 / TM65 / Airy Modified 1849 / Irish Grid}
\format{An object of class \code{character} of length 1.}
\usage{
crs.ni
}
\description{
Coordinate system: Cartesian 2D CS
Axes: easting, northing (E,N)
proj4js: "+proj=tmerc +lat_0=53.5 +lon_0=-8 +k=1.000035 +x_0=200000 +y_0=250000 +a=6377340.189 +b=6356034.447938534 +units=m +no_defs"
Bounds: [WGS84] ; [Projected]
Centroid: [WGS84] ; [Projected]
Unit: metre
Area of use: Ireland - onshore
Accuracy: 1m
}
\details{
sp::CRS(crs_code) to see the proj4js format for the specified epsg code
rgdal::make_EPSG() to build a table with code, name and proj4js format for all epsg codes
}
\keyword{datasets}
