% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{crs.wgs}
\alias{crs.wgs}
\title{Coordinate Reference Systems (CRS) in EPSG code for WGS84 / World Geodetic System 1984 / WGS84 / GPS Worlwide}
\format{
An object of class \code{character} of length 1.
}
\usage{
crs.wgs
}
\description{
Coordinate system: Ellipsoidal 2D CS
Axes: latitude, longitude
proj4js: "+init=epsg:4326 +proj=longlat +datum=WGS84 +no_defs +ellps=WGS84 +towgs84=0,0,0";
Bounds: NA
Centroid: NA
Unit: degree
Area of use: World
Accuracy: NA
}
\details{
sp::CRS(crs_code) to see the proj4js format for the specified epsg code
rgdal::make_EPSG() to build a table with code, name and proj4js format for all epsg codes
}
\keyword{datasets}
