% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{data_path}
\alias{data_path}
\title{Location of the UK datasets in the shared repository}
\format{
An object of class \code{character} of length 1.
}
\usage{
data_path
}
\description{
Location of the UK datasets in the shared repository
}
\keyword{datasets}
