% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{datauk_path}
\alias{datauk_path}
\title{Location of the geographic datasets in the shared repository}
\format{An object of class \code{character} of length 1.}
\usage{
datauk_path
}
\description{
Location of the geographic datasets in the shared repository
}
\keyword{datasets}
