% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{dataus_path}
\alias{dataus_path}
\title{Location of the US datasets in the shared repository}
\format{
An object of class \code{character} of length 1.
}
\usage{
dataus_path
}
\description{
Location of the US datasets in the shared repository
}
\keyword{datasets}
