% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databases.R
\name{dbm_do}
\alias{dbm_do}
\title{Perform an action on a local MySQL database}
\usage{
dbm_do(
  dname,
  action = "r",
  tname = NA,
  dts = NULL,
  strSQL = NA,
  trunc = TRUE,
  drop = FALSE
)
}
\arguments{
\item{dname}{the name of the database in which perform the action the table}

\item{action}{the action to be applied, among r (default), w, q, and s}

\item{tname}{the name of the table to be involved}

\item{dts}{the dataset to be uploaded if table has to be (over)written}

\item{strSQL}{the sql statement to get or send}

\item{trunc}{describe if the table has to be truncated before data being written}

\item{drop}{describe if the table has to be dropped before data being written}
}
\value{
A data.table, if action is 'r' or 'q'. None, otherwise
}
\description{
\code{dbm_do} allows to:
  - (r)ead a table
  - (w)rite a table
  - (q)uery a table
  - (s)end a query
}
\author{
Luca Valnegri, \email{l.valnegri@datamaps.co.uk}
}
