% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databases.R
\name{dbm_update}
\alias{dbm_update}
\title{Update an existing table in a local MySQL database based on a specified dataset}
\usage{
dbm_update(dbn, dts, update_sql, alter_sql = NA)
}
\arguments{
\item{dbn}{the name of the database that include the table to be updated}

\item{dts}{the dataset with the new values}

\item{update_sql}{the sql query that specifies the updating}

\item{alter_sql}{the optional action to be applied to the temp table before the updating (like setting a primary key or indices)}
}
\value{
None
}
\description{
Update an existing table in a local MySQL database based on a specified dataset
}
\author{
Luca Valnegri, \email{l.valnegri@datamaps.co.uk}
}
