% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geography.R
\name{dd_bbox}
\alias{dd_bbox}
\title{dd_bbox}
\usage{
dd_bbox(x_lon, y_lat, dst, in.miles = TRUE)
}
\arguments{
\item{x_lon}{double. The longitude of the centre point.}

\item{y_lat}{double. The latitude of the centre point.}

\item{dst}{double. The distance from the centre point.}

\item{in.miles}{logical (optional). If \code{TRUE} uses \emph{miles} as the units of \code{dist}, otherwise \emph{kilometres}.}
}
\value{
a matrix with two rows (the axis) and two columns (the limits)
}
\description{
Calculate a square or circle bounding box given a distance and a pair of coordinates
}
\references{
\url{http://janmatuschek.de/LatitudeLongitudeBoundingCoordinates}
}
\author{
Luca Valnegri, \email{l.valnegri@datamaps.co.uk}
}
