% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette.R
\name{dd_colour}
\alias{dd_colour}
\title{dd_colour}
\usage{
dd_colour(id, lbl, slc, full = TRUE)
}
\arguments{
\item{id}{id for the control}

\item{lbl}{label for the control}

\item{slc}{the colour scheme at start}

\item{full}{if \code{TRUE} returns the \emph{full} object, otherwise the \emph{minimal compact} version}
}
\value{
a \code{colorPickr} control
}
\description{
A tailored \code{colorPickr} control from the \code{shinyWidgets} package
}
\author{
Luca Valnegri, \email{l.valnegri@datamaps.co.uk}
}
