% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databases.R
\name{dd_db_connect}
\alias{dd_db_connect}
\title{Allows to connect to a MySQL dtabase server using the "group" credentials stored in the configuration file}
\usage{
dd_db_connect(x, dbn = NA)
}
\arguments{
\item{x}{the name of the group under which credentials are stored}

\item{dbn}{the optional name of the database to connect to, if it is not already present in the group credentials}
}
\value{
None
}
\description{
Allows to connect to a MySQL dtabase server using the "group" credentials stored in the configuration file
}
\author{
Luca Valnegri, \email{l.valnegri@datamaps.co.uk}
}
