% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geography.R
\name{dd_df2spat}
\alias{dd_df2spat}
\title{dd_conv2spat}
\usage{
dd_df2spat(
  x,
  cid = "id",
  clon = "Easting",
  clat = "Northing",
  output = "sf",
  all = FALSE,
  drop_coords = TRUE,
  out_names = c("x_lon", "y_lat"),
  crs.in = 27700,
  crs.out = 4326
)
}
\arguments{
\item{x}{a data.table with at least 3 columns: one for the id, the other two for the coordinates}

\item{cid}{the name of the id column}

\item{clon}{the name of the longitude or Easting column}

\item{clat}{the name of the latitude or Northing column}

\item{output}{specify which output to return: sf, sp, df, dt, map.}

\item{all}{when \code{FALSE}, only \code{id} and the coordinates are returned, otherwise all of \code{x} is returned}

\item{drop_coords}{when \code{FALSE}, only \code{id} and the coordinates are returned}

\item{out_names}{the name of the coordinates columns in the output table}

\item{crs.in}{the CRS of the input dataset}

\item{crs.out}{the CRS of the returning object}
}
\value{
depending on the choice of \code{output}, an object of type: sf, SpatialPointsDataFrame, dataframe, data.table, leaflet
}
\description{
Convert a data.table in a spatial object
}
\author{
Luca Valnegri, \email{l.valnegri@datamaps.co.uk}
}
