% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geography.R
\name{dd_do_pip}
\alias{dd_do_pip}
\title{dd_do_pip}
\usage{
dd_do_pip(
  x,
  y,
  cid = "id",
  clon = "x_lon",
  clat = "y_lat",
  pid = "id",
  pname = NA,
  output = "mgdf",
  verbose = FALSE,
  drop_void = TRUE,
  ...
)
}
\arguments{
\item{x}{a data.table with at least 3 columns, one for the id, the other two for the coordinates}

\item{y}{an \code{sf} object}

\item{cid}{the name of the column in the dataframe to be considered as <id>}

\item{clon}{the name of the column in the dataframe to be considered as longitude}

\item{clat}{the name of the column in the dataframe to be considered as latitude}

\item{pid}{the name of the column in the polygons to be considered as <id>}

\item{pname}{how to name the pid column in the output}

\item{output}{the type of output:
\code{df} returns only the two matching ids for points and polygons
\code{mgdf} merge the result of the operation with the input dataset \code{x}
\code{map} a \emph{leaflet} map}

\item{verbose}{if \code{TRUE}, annoying message on the steps of the process}

\item{drop_void}{if \code{TRUE} and \code{output} is \code{map}, delete all polygons without any included point before mapping}

\item{...}{Additional parameters to pass to the \code{basemap} function when output is \code{map}}
}
\value{
a data.table or a leaflet map
}
\description{
Calculates \emph{points in polygons}, returning an augmented data.table or a map
}
\author{
Luca Valnegri, \email{l.valnegri@datamaps.co.uk}
}
