% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_table.R
\name{dd_dt_update_col}
\alias{dd_dt_update_col}
\title{Update one column of a data.table based on a second lookup data.table}
\usage{
dd_dt_update_col(y, yk, ny, vy, nk = NULL, vk = NULL)
}
\arguments{
\item{y}{The data.table to be updated}

\item{yk}{The lookup data.table with \emph{ids} and \emph{values} to update in \code{y}}

\item{ny}{The name of the column in \code{y} to consider as \emph{ids}}

\item{vy}{The name of the column in \code{y} to consider as \emph{values}}

\item{nk}{The name of the column in \code{yk} to consider as \emph{ids}. If \code{NULL} it equates to \code{ny}}

\item{vk}{The name of the column in \code{yk} to consider as \emph{values}. If \code{NULL} it equates to \code{vy}}
}
\value{
A data.table
}
\description{
Update one column of a data.table based on a second lookup data.table
}
\author{
Luca Valnegri, \email{l.valnegri@datamaps.co.uk}
}
