% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geography.R
\name{dd_geocode}
\alias{dd_geocode}
\title{dd_geocode}
\usage{
dd_geocode(
  y,
  rgn,
  na = "address",
  nx = "x_lon",
  ny = "y_lat",
  update_address = FALSE
)
}
\arguments{
\item{y}{A data.table with an \emph{address} column}

\item{rgn}{The ISO2 2-chars string for the Country}

\item{na}{The name for the \emph{address} column}

\item{nx}{The name for the \emph{longitude} column (if it already exists it will be deleted)}

\item{ny}{The name for the \emph{latitude} column (if it already exists it will be deleted)}

\item{update_address}{if \code{TRUE}, the address column will be updated with the result from the Google Maps Engine}
}
\value{
A data.table
}
\description{
Geocode addresses in a data.table, adding longitude and latitude coordinates using the \verb{Google Maps Engine}
}
\author{
Luca Valnegri, \email{l.valnegri@datamaps.co.uk}
}
