% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postcodes.R
\name{dd_included_pcus}
\alias{dd_included_pcus}
\title{Given a code for a valid geographic area, return the set of all included PostCode Units}
\usage{
dd_included_pcus(x)
}
\arguments{
\item{x}{a code for a valid geographic area (see the table \code{zones})}
}
\value{
a data.table
}
\description{
Given a code for a valid geographic area, return the set of all included PostCode Units
}
\author{
Luca Valnegri, \email{l.valnegri@datamaps.co.uk}
}
