% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datetime.R
\name{dd_months}
\alias{dd_months}
\title{Create a dataset of ordered months, possibly by year(s)}
\usage{
dd_months(ys = NA, ye = NA)
}
\arguments{
\item{ys}{the start year, or the only year if \code{ye} is missing. If missing, the function returns only the months.}

\item{ye}{the final year}
}
\value{
a data.table
}
\description{
Create a dataset of ordered months, possibly by year(s)
}
\author{
Luca Valnegri, \email{l.valnegri@datamaps.co.uk}
}
