% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postcodes.R
\name{dd_nearest_pcus}
\alias{dd_nearest_pcus}
\title{Given a Postcode Unit and a distance in metres, return the set of PostCode Units within that distance}
\usage{
dd_nearest_pcus(x, d)
}
\arguments{
\item{x}{a PostCode Unit in any valid form}

\item{d}{a distance in metres}
}
\value{
a data.table
}
\description{
Given a Postcode Unit and a distance in metres, return the set of PostCode Units within that distance
}
\author{
Luca Valnegri, \email{l.valnegri@datamaps.co.uk}
}
