% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feat_eng.R
\name{dd_ohe}
\alias{dd_ohe}
\title{dd_ohe}
\usage{
dd_ohe(x, add_is = TRUE)
}
\arguments{
\item{x}{A vector of values for an attribute}

\item{add_is}{if \code{TRUE}, add \code{is_} in front of each value as the column name}
}
\value{
A data.table with as many columns as the unique values in the input.
}
\description{
One Hot Encoding
}
\author{
Luca Valnegri, \email{l.valnegri@datamaps.co.uk}
}
