% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette.R
\name{dd_palette}
\alias{dd_palette}
\title{dd_palette}
\usage{
dd_palette(id, lbl = "PALETTE:", sel = "Red-Yellow-Blue")
}
\arguments{
\item{id}{id for the  control}

\item{lbl}{label for the control}

\item{sel}{colour scheme at start}
}
\value{
a \code{palettePicker} control
}
\description{
A tailored \code{palettePicker} control from the \code{esquisse} package
}
\author{
Luca Valnegri, \email{l.valnegri@datamaps.co.uk}
}
