% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postcodes.R
\name{dd_pcu_coords}
\alias{dd_pcu_coords}
\title{Extract the coordinates of a postcode}
\usage{
dd_pcu_coords(x)
}
\arguments{
\item{x}{a PostCode Unit in any valid form}
}
\value{
A data.table
}
\description{
Extract the coordinates of a postcode
}
\author{
Luca Valnegri, \email{l.valnegri@datamaps.co.uk}
}
