% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postcodes.R
\name{dd_pcu_from_address}
\alias{dd_pcu_from_address}
\title{Given a data.table with a column containing allegedly addresses, it returns a column of PostCode Units}
\usage{
dd_pcu_from_address(x)
}
\arguments{
\item{x}{A data.table containing a column with texts that should correspond to addresses}
}
\value{
A character
}
\description{
Given a data.table with a column containing allegedly addresses, it returns a column of PostCode Units
}
\author{
Luca Valnegri, \email{l.valnegri@datamaps.co.uk}
}
