% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postcodes.R
\name{dd_pcu_from_coords}
\alias{dd_pcu_from_coords}
\title{Given a pair of coordinates, returns the nearest PostCode Unit (centroid)}
\usage{
dd_pcu_from_coords(x_lon, y_lat)
}
\arguments{
\item{y_lat}{}
}
\value{
A data.table
}
\description{
Given a pair of coordinates, returns the nearest PostCode Unit (centroid)
}
\author{
Luca Valnegri, \email{l.valnegri@datamaps.co.uk}
}
