% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boundaries.R
\name{dd_save_bnd}
\alias{dd_save_bnd}
\title{dd_save_bnd}
\usage{
dd_save_bnd(bnd, bname, bpath, fmt = "q", fmt2path = FALSE, pct = NULL)
}
\arguments{
\item{bnd}{The \code{sf} object to be saved}

\item{bname}{The name of the output file}

\item{bpath}{The \emph{root} output folder where to store the file(s)}

\item{fmt}{The output format(s) of the file
One or more in:
- \code{s} for \emph{shapefile} (folders uses \code{shp})
- \code{j} for \emph{geojson} (\code{gjs})
- \code{k} for \emph{KML}     (\code{kml})
- \code{q} for \code{qs} ($R$ serialization) (\code{qs})
- \code{r} for \code{RDS} (native $R$ format) (\code{rds})}

\item{fmt2path}{Add the extension of the output format (or an acronym) to the output folder \code{bpath}}

\item{pct}{\verb{NULL`` or a 2-char string identifying the percentage of simplification. One in: '00', '10', '20', '30', '40', '50' ('00' stand for *original size*) IF not }NULL\verb{, the string will be be added to the output folder (after }fmt`)}
}
\value{
none
}
\description{
Save an \code{sf} object as one or more in: \emph{shapefile}, \emph{geojson}, \emph{KML}, \code{qs}, \code{RDS}.
}
\author{
Luca Valnegri, \email{l.valnegri@datamaps.co.uk}
}
