% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{dd_tab_pct}
\alias{dd_tab_pct}
\title{Build a simple frequency table with relative/percentage and cumulative}
\usage{
dd_tab_pct(
  x,
  use_na = FALSE,
  ord_freq = FALSE,
  ord_desc = FALSE,
  add_pct = FALSE,
  mlt_pct = 100,
  dgt_pct = 2,
  add_cum = NA,
  col_names = NULL
)
}
\arguments{
\item{x}{a vector of values to be tabulated}

\item{use_na}{if TRUE the table will include missing values as a separate value/category}

\item{ord_freq}{if TRUE the table will be ordered by frequency instead of attributes}

\item{ord_desc}{if TRUE the table will be ordered in decreasing instead of increasing}

\item{add_pct}{to add relative/percentage frequency}

\item{mlt_pct}{an integer multiplier for the "percentage" column}

\item{dgt_pct}{the decimal digits to keep for the "percentage" column}

\item{add_cum}{a character string to indicate to add cumulative frequency ('f') and/or percentage ('p')}

\item{col_names}{a character vector of the required number of elements for the column names in the output dataset.
By default, they are: \code{c('X', 'freq', 'pct', 'cum_freq', 'cum_pct')}}
}
\value{
A data.table
}
\description{
Build a simple frequency table with relative/percentage and cumulative
}
\author{
Luca Valnegri, \email{l.valnegri@datamaps.co.uk}
}
