% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmpkg.funs.R
\docType{package}
\name{dmpkg.funs}
\alias{dmpkg.funs}
\title{dmpkg.funs: a bunch of functions to help Data Sourcing, Cleaning and Wrangling for data projects}
\description{
The dmpkg.funs package provides the following categories of important functions:
 - connection and querying to local MySQL and main SQL Server databases
 - cleaning, recoding, saving, and querying R datasets
 - quick summaries and mapping
}
