% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geography.R
\name{do_pip}
\alias{do_pip}
\title{Calculates points in polygons, returning an augmented data.table or a map}
\usage{
do_pip(
  x,
  y,
  cid = "id",
  clon = "x_lon",
  clat = "y_lat",
  pid = "id",
  pname = NA,
  output = "mgdf",
  verbose = FALSE,
  drop_void = TRUE,
  ...
)
}
\arguments{
\item{x}{a data.table with at least 3 columns, one for the id, the other two for the coordinates}

\item{y}{a SpatialPolygonsDataFrame}

\item{cid}{the name of the column in the dataframe to be considered as <id>}

\item{clon}{the name of the column in the dataframe to be considered as longitude}

\item{clat}{the name of the column in the dataframe to be considered as latitude}

\item{pid}{the name of the column in the polygons to be considered as <id>}

\item{pname}{how to name the pid column in the output}

\item{output}{the type of output: 'df' returns only the two matching ids for points and polygons, 'mgdf' merge the result of the operation with the input df, <map> a leaflet map}

\item{verbose}{if TRUE, annoying message on the steps of the process}

\item{drop_void}{if TRUE and output is <map>, delete all polygons without any included point before mapping}

\item{...}{Additional parameters to pass to the <basemap> function if output is <map>}
}
\value{
a data.table or a leaflet map
}
\description{
Calculates points in polygons, returning an augmented data.table or a map
}
\author{
Luca Valnegri, \email{l.valnegri@datamaps.co.uk}
}
