% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geography.R
\name{do_pip_uk}
\alias{do_pip_uk}
\title{Calculates points in polygons using default UK polygons, returning an augmented data.table}
\usage{
do_pip_uk(x, areatype = "OA", ...)
}
\arguments{
\item{x}{a data.table with at least 3 columns, one for the id, the other two for the geographic coordinates}

\item{areatype}{the reference area to operate on for the point in polygon process}

\item{...}{Additional parameters to pass to the <do_pip> function, and/or the <basemap> function if "output" is <map>}
}
\value{
a data.table
}
\description{
Calculates points in polygons using default UK polygons, returning an augmented data.table
}
\author{
Luca Valnegri, \email{l.valnegri@datamaps.co.uk}
}
