% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{down_shp}
\alias{down_shp}
\title{down_shp}
\usage{
down_shp(x, fn = NA, lst = FALSE)
}
\arguments{
\item{x}{a url for a zipped file containing one or more shapefiles}

\item{fn}{a specific name to extract and convert}

\item{lst}{only prints the names of the included files}
}
\value{
an \code{sf} object
}
\description{
Download a zipped \emph{shapefile} and returns it as an \code{sf} object.
If the zip file contains multiple shapefiles, a specific file name can be passed, or the first will be returned.
}
\author{
Luca Valnegri, \email{l.valnegri@datamaps.co.uk}
}
