% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{down_unz_ren_shp}
\alias{down_unz_ren_shp}
\title{Download a zipped shapefile, unzip it in a specified folder, then rename extracted files with a common name}
\usage{
down_unz_ren_shp(x, d, prfx = NULL)
}
\arguments{
\item{x}{a url for a zipped file}

\item{d}{the directory where the unzipped files should be saved}

\item{prfx}{the common name for the unzipped files}
}
\value{
none
}
\description{
Download a zipped shapefile, unzip it in a specified folder, then rename extracted files with a common name
}
\author{
Luca Valnegri, \email{l.valnegri@datamaps.co.uk}
}
