% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangling.R
\name{dt_update}
\alias{dt_update}
\title{Update a column in a data.table (partially or totally) using the values of another column}
\usage{
dt_update(dt, lkp, to_factor = FALSE)
}
\arguments{
\item{dt}{a data.table}

\item{lkp}{a mapping between two column of dt, the names must reflect the names in *dt*}

\item{to_factor}{if TRUE convert the updated column as factor}
}
\value{
a data.table
}
\description{
Update a column in a data.table (partially or totally) using the values of another column
}
\author{
Luca Valnegri, \email{l.valnegri@datamaps.co.uk}
}
