% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dunzip}
\alias{dunzip}
\title{dunzip}
\usage{
dunzip(
  url,
  fname = NA,
  out_path = NA,
  ext = "csv",
  shp = FALSE,
  bndid = NA,
  btransform = TRUE,
  crs = crs.wgs,
  bsimplify = FALSE,
  spct = 0.05
)
}
\arguments{
\item{url}{the url of the zip file to be downloaded}

\item{shp}{if TRUE, the zip file contains shapefiles; if FALSE a file with the extension specified in <ext>}

\item{bndid}{if not NA, the name(s) of the column(s) to be carried over. If only one column, that column become the new id of the object}

\item{btransform}{if not NA, the polygons}

\item{crs}{the new Coordinate System for the polygons}

\item{spct}{}
}
\value{
None, a data.table, or a SpatialPolygons{DataFrame}
}
\description{
Download zip files and save inner files depending on some content
}
\author{
Luca Valnegri, \email{l.valnegri@datamaps.co.uk}
}
