% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{ext_path}
\alias{ext_path}
\title{Root of any external source data}
\format{
An object of class \code{character} of length 1.
}
\usage{
ext_path
}
\description{
Root of any external source data
}
\keyword{datasets}
