% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{face.types}
\alias{face.types}
\title{List of available font types}
\format{An object of class \code{character} of length 4.}
\usage{
face.types
}
\description{
List prepared to be used as choices in Shiny input controls
}
\keyword{datasets}
