% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps_leaflet.R
\docType{data}
\name{faPlugin}
\alias{faPlugin}
\title{faPlugin}
\format{
An object of class \code{html_dependency} of length 10.
}
\usage{
faPlugin
}
\description{
Defines a dependency to allow the use of *fontawesome pro* icons pro in `leaflet` maps
}
\references{
\url{https://fontawesome.com/v6/docs/web/style/styling}.
}
\keyword{datasets}
