% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{fcstypes.lst}
\alias{fcstypes.lst}
\title{List of available font types}
\format{
An object of class \code{character} of length 4.
}
\usage{
fcstypes.lst
}
\description{
List of available font types
}
\keyword{datasets}
