% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geography.R
\name{fix_hole}
\alias{fix_hole}
\title{Rewrite a polygon inside a SpatialPolygonsDataframe cutting out some of its interior equal to another of its polygons}
\usage{
fix_hole(bnd, outer, hole)
}
\arguments{
\item{bnd}{the SpatialPolygonsDataframe object of the change}

\item{outer}{the id of the parent polygon}

\item{hole}{the id of the child (hole) polygon}
}
\value{
a SpatialPolygonsDataFrame
}
\description{
Rewrite a polygon inside a SpatialPolygonsDataframe cutting out some of its interior equal to another of its polygons
}
\author{
Luca Valnegri, \email{l.valnegri@datamaps.co.uk}
}
