% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geography.R
\name{fix_holes}
\alias{fix_holes}
\title{Rewrite a list of polygons inside a SpatialPolygonsDataframe cutting out some of its interior equal to another of its polygons}
\usage{
fix_holes(bnd, ids)
}
\arguments{
\item{bnd}{the SpatialPolygonsDataframe object of the change}

\item{ids}{a 2-cols dataframe, the first column with the ids of the outer polygons, the second the ids of the corresponding holes}
}
\value{
a SpatialPolygonsDataFrame
}
\description{
Rewrite a list of polygons inside a SpatialPolygonsDataframe cutting out some of its interior equal to another of its polygons
}
\author{
Luca Valnegri, \email{l.valnegri@datamaps.co.uk}
}
