% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{geoit_path}
\alias{geoit_path}
\title{Location of the geographic Italian datasets in the shared repository}
\format{
An object of class \code{character} of length 1.
}
\usage{
geoit_path
}
\description{
Location of the geographic Italian datasets in the shared repository
}
\keyword{datasets}
