% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{geouk_path}
\alias{geouk_path}
\title{Location of the geographic UK datasets in the shared repository}
\format{An object of class \code{character} of length 1.}
\usage{
geouk_path
}
\description{
Location of the geographic UK datasets in the shared repository
}
\keyword{datasets}
