% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geography.R
\name{get_area_code}
\alias{get_area_code}
\title{Return all entries in the locations list including or exactly with the specified string}
\usage{
get_area_code(x, tpe = NA, exact = FALSE)
}
\arguments{
\item{x}{the string to look for}

\item{tpe}{the location type. If NA he search is done on all the records}

\item{exact}{if FALSE uses a regex search}
}
\value{
a data.table with two or three columns, depending on the tpe parameter
}
\description{
Return all entries in the locations list including or exactly with the specified string
}
\author{
Luca Valnegri, \email{l.valnegri@datamaps.co.uk}
}
