% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postcodes.R
\name{get_areas_ovelapping}
\alias{get_areas_ovelapping}
\title{Find overlapping areas created when merging children areas using the postcodes table. If asked, updates the table itself}
\usage{
get_areas_ovelapping(parent, child, dts = NULL, update = FALSE)
}
\arguments{
\item{parent}{The location type of the higher level area}

\item{child}{The location type of the lower level area}

\item{dts}{If TRUE, it return the postcodes tables}

\item{update}{If TRUE, it updates the saved or passed postcodes table, then returns it.
Otherwise, it returns the table of overlapping children}
}
\value{
A data table of overlapping areas, or the updated postcodes table
}
\description{
Find overlapping areas created when merging children areas using the postcodes table. If asked, updates the table itself
}
\author{
Luca Valnegri, \email{l.valnegri@datamaps.co.uk}
}
