% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_csv_zip}
\alias{get_csv_zip}
\title{Download a zip file and extract a csv file therein.
If multiple csv files are included in the zip file, only the biggest one is returned, unless a name is given.}
\usage{
get_csv_zip(url, fname = NULL, cols = NULL, coln = NULL)
}
\arguments{
\item{url}{the url of the zip file to be downloaded}

\item{fname}{the name of the csv to be read. When NULL, the biggest , or the only, file included.}

\item{cols}{to restrict selection of columns}

\item{coln}{to change column names}
}
\value{
a data.table
}
\description{
Download a zip file and extract a csv file therein.
If multiple csv files are included in the zip file, only the biggest one is returned, unless a name is given.
}
\author{
Luca Valnegri, \email{l.valnegri@datamaps.co.uk}
}
