% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{get_fxb_labels}
\alias{get_fxb_labels}
\title{Determines labels for (n + 1) bins given a series of n breaks or a convenient building method}
\usage{
get_fxb_labels(y, dec.fig = 1, del_signs = TRUE)
}
\arguments{
\item{y}{a vector describing}

\item{beta}{<description of beta>}
}
\value{
a data.table with limits and labels
}
\description{
Insert (longer) >>> DESCRIPTION <<< here ...
}
\author{
Luca Valnegri, \email{l.valnegri@datamaps.co.uk}
}
