% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{get_map_legend}
\alias{get_map_legend}
\title{Build a legend to be used in a leaflet thematic map}
\usage{
get_map_legend(mtc, brks, dec.fig = 2, del_signs = TRUE)
}
\arguments{
\item{alpha}{<description of alpha>}

\item{beta}{<description of beta>}
}
\value{
describe what the function is giving back to the user (insert "None" if there's no value returned)
}
\description{
Build a legend to be used in a leaflet thematic map
}
\examples{

}
\references{
leaflet, maps
}
\author{
Luca Valnegri, \email{l.valnegri@datamaps.co.uk}
}
\keyword{}
