% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_num_sfx}
\alias{get_num_sfx}
\title{Insert >>> TITLE <<< here ...}
\usage{
get_num_sfx(x)
}
\arguments{
\item{alpha}{<description of alpha>}

\item{beta}{<description of beta>}
}
\value{
describe what the function is giving back to the user (insert "None" if there's no value returned)
}
\description{
Insert (longer) >>> DESCRIPTION <<< here ...
}
\examples{

}
\references{
leaflet, maps
}
\seealso{
\code{\link{...}}
}
\author{
Luca Valnegri, \email{l.valnegri@datamaps.co.uk}
}
\keyword{}
