% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postcodes.R
\name{get_postcode_coords}
\alias{get_postcode_coords}
\title{Extraxt the coordinates of a postcode}
\usage{
get_postcode_coords(postcode)
}
\arguments{
\item{postcode}{}
}
\value{
a data.table
}
\description{
Extraxt the coordinates of a postcode
}
\author{
Luca Valnegri, \email{l.valnegri@datamaps.co.uk}
}
