% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postcodes.R
\name{get_postcode_neighbors}
\alias{get_postcode_neighbors}
\title{Calculate the set of postcodes included in a bounding box or circle with a specified postcode unit as centroid}
\usage{
get_postcode_neighbors(
  postcode,
  distance = 0.5,
  circle = TRUE,
  in.miles = TRUE,
  active_only = TRUE
)
}
\arguments{
\item{active_only}{}
}
\value{
a data.table
}
\description{
Calculate the set of postcodes included in a bounding box or circle with a specified postcode unit as centroid
}
\author{
Luca Valnegri, \email{l.valnegri@datamaps.co.uk}
}
