% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postcodes.R
\name{get_postcodes_area}
\alias{get_postcodes_area}
\title{Returns the set of postcodes included in a specified geographical area}
\usage{
get_postcodes_area(ids, active_only = TRUE)
}
\arguments{
\item{active_only}{If TRUE, only active postcodes are returned.
If FALSE, all postcodes are returned, with an additional <is_active> column to determine if a postcode is active or terminated}

\item{id}{The ONS code of the location we want postcodes from}
}
\value{
a data.table
}
\description{
Returns the set of postcodes included in a specified geographical area
}
\author{
Luca Valnegri, \email{l.valnegri@datamaps.co.uk}
}
