% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot.R
\name{gg.to.ggiraph}
\alias{gg.to.ggiraph}
\title{Convert a \code{ggplot} into its corresponding \code{htmlwidgets} interactive plot using the \code{ggiraph} extension}
\usage{
gg.to.ggiraph(p, sel.type = "single", gg.width = 0.8)
}
\arguments{
\item{sel.type}{}
}
\value{
a ggiraph ggplot
}
\description{
Convert a \code{ggplot} into its corresponding \code{htmlwidgets} interactive plot using the \code{ggiraph} extension
}
\author{
Luca Valnegri, \email{l.valnegri@datamaps.co.uk}
}
