% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps_leaflet.R
\docType{data}
\name{hlt.options}
\alias{hlt.options}
\title{hlt.options}
\format{
An object of class \code{list} of length 5.
}
\usage{
hlt.options
}
\description{
Generic options for the border highlight when hovering polygons in \code{leaflet} maps
}
\keyword{datasets}
