% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangling.R
\name{insert_column}
\alias{insert_column}
\title{Reorder the columns in a data.table so that a column, possibly renamed, is listed after another column}
\usage{
insert_column(dt, colname, after, rename = NULL)
}
\arguments{
\item{dt}{a data.table}

\item{colname}{the name of the column to be moved}

\item{after}{the name of the column after which \em{colname} the will be positioned}

\item{rename}{the name to substitute for the moved column}
}
\value{
a data.table
}
\description{
Reorder the columns in a data.table so that a column, possibly renamed, is listed after another column
}
\author{
Luca Valnegri, \email{l.valnegri@datamaps.co.uk}
}
