% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps_leaflet.R
\docType{data}
\name{lbl.options}
\alias{lbl.options}
\title{lbl.options}
\format{
An object of class \code{list} of length 11.
}
\usage{
lbl.options
}
\description{
Generic options for the labels appearing when hovering polygons in `leaflet` maps
}
\keyword{datasets}
