% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot.R
\name{lbl.plt.rotation}
\alias{lbl.plt.rotation}
\title{Return correct spacing for axis labels rotation in \code{ggplot2} charts to use in \code{my_ggtheme}}
\usage{
lbl.plt.rotation(angle, position = "x")
}
\value{
a ggplot
}
\description{
Return correct spacing for axis labels rotation in \code{ggplot2} charts to use in \code{my_ggtheme}
}
\author{
Luca Valnegri, \email{l.valnegri@datamaps.co.uk}
}
