% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{line.types}
\alias{line.types}
\title{List of available line types}
\format{An object of class \code{character} of length 6.}
\usage{
line.types
}
\description{
List of available line types
}
\keyword{datasets}
