% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{lnstypes.lst}
\alias{lnstypes.lst}
\title{List of available line types}
\format{
An object of class \code{character} of length 6.
}
\usage{
lnstypes.lst
}
\description{
List of available line types
}
\keyword{datasets}
