% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postcodes.R
\name{lookup_postcodes_shp}
\alias{lookup_postcodes_shp}
\title{Return the code mapping between postcodes and a shapefile boundaries online (from the ONS geoportal) referring to a specified location type}
\usage{
lookup_postcodes_shp(furl, onsid, tpe)
}
\arguments{
\item{furl}{the url of the shapefile}

\item{onsid}{the name of the column in the shapefile that represents the}

\item{tpe}{}
}
\value{
a data.table with two columns, postcode and location type
}
\description{
Return the code mapping between postcodes and a shapefile boundaries online (from the ONS geoportal) referring to a specified location type
}
\examples{
\dontrun{
}

}
\author{
Luca Valnegri, \email{l.valnegri@datamaps.co.uk}
}
