% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postcodes.R
\name{map_postcodes_area}
\alias{map_postcodes_area}
\title{Return a map of the set of postcodes included in a bounding box or circle with a specified postcode unit as centroid}
\usage{
map_postcodes_area(
  ids,
  active_only = FALSE,
  ch = TRUE,
  oa = TRUE,
  tiles = tiles.lst,
  use_icons = FALSE
)
}
\arguments{
\item{active_only}{If TRUE, only active postcodes are mapped.
If FALSE, all postcodes are mapped, divided in two separate sub-maps as <active> and <terminated>}

\item{id}{The ONS code of the location we want the map from}
}
\value{
a leaflet map
}
\description{
Return a map of the set of postcodes included in a bounding box or circle with a specified postcode unit as centroid
}
\author{
Luca Valnegri, \email{l.valnegri@datamaps.co.uk}
}
