% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postcodes.R
\name{map_postcodes_areaname}
\alias{map_postcodes_areaname}
\title{Return a map of a location passed using a string instead of a code}
\usage{
map_postcodes_areaname(x, tpe = NA, exact = FALSE, ...)
}
\arguments{
\item{...}{accept all parameters used in <map_postcodes_area> apart from the id}
}
\value{
a leaflet map
}
\description{
Return a map of a location passed using a string instead of a code
}
\examples{
\dontrun{
}

}
\author{
Luca Valnegri, \email{l.valnegri@datamaps.co.uk}
}
