% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps_leaflet.R, R/palette.R
\docType{data}
\name{marker_colours}
\alias{marker_colours}
\title{marker_colours}
\format{
An object of class \code{character} of length 16.

An object of class \code{character} of length 16.
}
\usage{
marker_colours

marker_colours
}
\description{
List of the available colours to fill \emph{Awesome Markers} in \code{leaflet} maps

List of the available colours to fill \emph{Awesome Markers} in \code{leaflet} maps
}
\references{
\url{https://rstudio.github.io/leaflet/markers.html}

\url{https://rstudio.github.io/leaflet/markers.html}
}
\keyword{datasets}
