% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{mnths}
\alias{mnths}
\title{list of months in their natural order}
\format{
An object of class \code{data.frame} with 36 rows and 2 columns.
}
\usage{
mnths
}
\description{
list of months in their natural order
}
\keyword{datasets}
