% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osrm.R
\name{mostra_percorso}
\alias{mostra_percorso}
\title{Crea e visualizza il percorso piu breve fra due punti, tenendo presente un profilo di locomozione}
\usage{
mostra_percorso(ps, pe, prf = 1, tessera = tiles.lst[[1]])
}
\arguments{
\item{ps}{coordinate del punto di partenza}

\item{pe}{coordinate del punto di arrivo}

\item{prf}{codice del profilo: 1-auto, 2-cammino, 3-bici}

\item{tessera}{tessera mappa di background (si veda l'elenco \code{tiles.lst})}
}
\value{
una mappa \emph{leaflet}
}
\description{
Crea e visualizza il percorso piu breve fra due punti, tenendo presente un profilo di locomozione
}
\author{
Luca Valnegri, \email{l.valnegri@datamaps.co.uk}
}
