% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps_leaflet.R
\name{mrkr_cls_lims}
\alias{mrkr_cls_lims}
\title{mrkr_cls_lims}
\usage{
mrkr_cls_lims(n = 1)
}
\arguments{
\item{n}{proportional factor to alter the four included fixed breaks: 0, 20, 100, 200}
}
\description{
Writes \code{JS} function to redefine \emph{Marker Clusters} in \code{leaflet} maps using five bins
}
\references{
\url{https://github.com/Leaflet/Leaflet.markercluster}.
}
\author{
Luca Valnegri, \email{l.valnegri@datamaps.co.uk}
}
