% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot.R
\name{my_ggtheme}
\alias{my_ggtheme}
\title{Global style for \code{ggplot2} charts}
\usage{
my_ggtheme(
  g,
  xaxis.draw = FALSE,
  yaxis.draw = FALSE,
  axis.draw = FALSE,
  ticks.draw = FALSE,
  axis.colour = "black",
  axis.size = 0.1,
  hgrid.draw = FALSE,
  vgrid.draw = FALSE,
  grids.colour = "black",
  grids.size = 0.1,
  grids.type = "dotted",
  labels.rotation = c(45, 0),
  labels.rotate = FALSE,
  font.size = 6,
  ttl.font.size.mult = 1.2,
  ttl.face = "bold",
  font.family = "Arial",
  bkg.colour = "white",
  plot.border = FALSE,
  legend.pos = "bottom"
)
}
\arguments{
\item{g}{a ggplot}
}
\value{
a ggplot
}
\description{
Global style for \code{ggplot2} charts
}
\author{
Luca Valnegri, \email{l.valnegri@datamaps.co.uk}
}
