% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{nchar_int}
\alias{nchar_int}
\title{nchar_int}
\usage{
nchar_int(x)
}
\arguments{
\item{x}{an integer}
}
\value{
an integer
}
\description{
Return the length in character of an integer (works also for very long integer)
}
\author{
Luca Valnegri, \email{l.valnegri@datamaps.co.uk}
}
