% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{output_areas}
\alias{output_areas}
\title{UK Geography Areas Lookup}
\format{
A data frame with nine variables:
\describe{
    \item{\code{OA}}{Output Area}
    \item{\code{LSOA}}{Output Area}
    \item{\code{MSOA}}{Output Area}
    \item{\code{LAD}}{}
    \item{\code{CTY}}{}
    \item{\code{RGN}}{}
    \item{\code{CTRY}}{}
    \item{\code{}}{}
}

For further details, see \url{https://geoportal.statistics.gov.uk/search?collection=Dataset&sort=-modified&tags=PRD_RGC}
}
\usage{
output_areas
}
\description{
A mapping between Output Areas and many other geographic areas in the United Kingdom
}
\keyword{datasets}
