% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette.R
\docType{data}
\name{palettes.lst.pkr}
\alias{palettes.lst.pkr}
\title{palettes.lst.pkr}
\format{
An object of class \code{list} of length 3.
}
\usage{
palettes.lst.pkr
}
\value{
a list
}
\description{
A named nested list of colour schemas to be used in Shiny controls
}
\author{
Luca Valnegri, \email{l.valnegri@datamaps.co.uk}
}
\keyword{datasets}
