% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{pntshapes.lst}
\alias{pntshapes.lst}
\title{List of available shapes with outline colour}
\format{
An object of class \code{numeric} of length 5.
}
\usage{
pntshapes.lst
}
\description{
List of available shapes with outline colour
}
\keyword{datasets}
