% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{point.shapes}
\alias{point.shapes}
\title{List of available shapes with outline colour}
\format{An object of class \code{numeric} of length 5.}
\usage{
point.shapes
}
\description{
List prepared to be used as choices in Shiny input controls
}
\keyword{datasets}
