% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popifun.R
\docType{package}
\name{popiFun}
\alias{popiFun}
\alias{popiFun-package}
\title{popiFun: a bunch of functions to help Data Sourcing, Cleaning and Wrangling for data projects}
\description{
The popiFun package provides the following categories of important functions:
 - connection and querying to local MySQL and main SQL Server databases
 - cleaning, recoding, saving, and querying R datasets
 - quick summaries and mapping
}
