% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{pos.lst}
\alias{pos.lst}
\title{List of available positions on the screen for objects that stays in a corner}
\format{
An object of class \code{character} of length 4.
}
\usage{
pos.lst
}
\description{
List of available positions on the screen for objects that stays in a corner
}
\keyword{datasets}
