% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fst.R
\name{read_fst_idx}
\alias{read_fst_idx}
\title{Read a (partial) dataset from an fst indexed file based on a set of values}
\usage{
read_fst_idx(fname, ref)
}
\arguments{
\item{fname}{the name of the fst complete of its path}

\item{values}{the set of values pertaining to the column that makes up the index}
}
\value{
A data.table
}
\description{
Read a (partial) dataset from an fst indexed file based on a set of values
}
\author{
Luca Valnegri, \email{l.valnegri@datamaps.co.uk}
}
