% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web.R
\name{read_html_ns}
\alias{read_html_ns}
\title{read_html_ns}
\usage{
read_html_ns(x, el = NULL)
}
\arguments{
\item{x}{The url of the file to download}

\item{el}{If not \code{NULL}, add an expression to retrieve the elements involved with the selector CSS \code{el}}
}
\value{
An \code{xml_document} or an \code{xml_nodeset} depending ont the value of the parameter \code{el}.
}
\description{
Read an HTML file when a website is not certified as secure, or self-certified
}
\author{
Luca Valnegri, \email{l.valnegri@datamaps.co.uk}
}
