% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databases.R
\name{read_mysql_cnf}
\alias{read_mysql_cnf}
\title{Read MySQL configuration file for "group" access}
\usage{
read_mysql_cnf(x)
}
\arguments{
\item{x}{the name of the group under which credentials are stored}
}
\value{
None
}
\description{
Read MySQL configuration file for "group" access
}
\author{
Luca Valnegri, \email{l.valnegri@datamaps.co.uk}
}
