% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databases.R
\name{rename_db}
\alias{rename_db}
\title{Allows to rename or copy a local MySQL database}
\usage{
rename_db(old_db, new_db, create_new = TRUE, drop_old = TRUE)
}
\arguments{
\item{old_db}{The name of the existing database}

\item{new_db}{The name of the database to be the new container}

\item{create_new}{IF TRUE any existing database called old_db will be dropped and created as new}

\item{drop_old}{If TRUE (the default) the old database will be dropped.
When FALSE, the database will be kept with all its tablesall}
}
\value{
None
}
\description{
Allows to rename or copy a local MySQL database
}
\author{
Luca Valnegri, \email{l.valnegri@datamaps.co.uk}
}
