% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangling.R
\name{reorder_columns}
\alias{reorder_columns}
\title{Reorder the columns in a data.table so that one or more columns, possibly renamed, are listed just after another column}
\usage{
reorder_columns(dt, cname, after, rename = NULL)
}
\arguments{
\item{dt}{a data.table}

\item{cname}{the name(s) of the column(s) to be moved}

\item{after}{the name of the column after which `cname` will be positioned}

\item{rename}{the name(s) to substitute for the moved column(s), of the same length of `cname`}
}
\value{
a data.table
}
\description{
Reorder the columns in a data.table so that one or more columns, possibly renamed, are listed just after another column
}
\author{
Luca Valnegri, \email{l.valnegri@datamaps.co.uk}
}
