% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rfuns.R
\docType{package}
\name{Rfuns}
\alias{Rfuns}
\title{Rfuns: a bunch of functions to help Data Sourcing, Cleaning and Wrangling for data projects}
\description{
The \code{Rfuns} package provides the following categories of important functions:
\itemize{
\item connection and querying to local MySQL and main SQL Server databases
\item cleaning, recoding, saving, and querying R datasets
\item quick summaries and mapping
}
}
