% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geography.R
\name{save_bnd}
\alias{save_bnd}
\title{Save a SpatialPolygonsDataframe either/both as a shapefile and/or in RDS format}
\usage{
save_bnd(bnd, bname, shp = TRUE, rds = TRUE, bpath = bnduk_path, pct = "00")
}
\arguments{
\item{bnd}{the SpatialPolygonsDataframe to be saved}

\item{bname}{the name of the output file}

\item{shp}{if true, save as shapefile}

\item{rds}{if true, save in RDS format}

\item{bpath}{the "root" directory where to store the  file}

\item{pct}{a 2-char string identifying the percentage of simplification: must be '05', '10', '20', '30', '40', '50'. '00' stand for "original size"}
}
\value{
none
}
\description{
Save a SpatialPolygonsDataframe either/both as a shapefile and/or in RDS format
}
\author{
Luca Valnegri, \email{l.valnegri@datamaps.co.uk}
}
