% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boundaries.R
\name{save_bnd_folder}
\alias{save_bnd_folder}
\title{save_bnd_folder}
\usage{
save_bnd_folder(x)
}
\arguments{
\item{bname}{The name of the output file}

\item{bpath}{The \emph{root} output folder where to store the file(s)

Build the output folder for the function \code{save_bnd}}
}
\description{
save_bnd_folder
}
