% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_dts_pkg.R
\name{save_dts_pkg}
\alias{save_dts_pkg}
\title{save_dts_pkg}
\usage{
save_dts_pkg(
  y,
  fn,
  fst_path = NULL,
  fst_cols = NULL,
  as_rdb = TRUE,
  dbn = NULL,
  tbn = fn,
  as_rda = TRUE,
  csv_in_pkg = TRUE,
  csv2zip = FALSE
)
}
\arguments{
\item{y}{a dataset}

\item{fn}{a common name for all the file formats}

\item{fst_path}{the (optional) path where to save the file in \code{fst} format (usually inside the \emph{public} repository)}

\item{fst_cols}{When not \code{NULL}, save the dataset as an indexed \code{fst} file}

\item{as_rdb}{if \code{TRUE}, save the dataset in the \code{dbn} database.}

\item{dbn}{the name of the database to be used when \code{as_rdb} is \code{TRUE}}

\item{tbn}{The name for the database table where to store the dataset (by default equals \code{fn}).
The table should be conveniently created beforehand.
The current table will be truncated, so that existing data will be deleted}

\item{as_rda}{if \code{TRUE}, and when inside a package project, save the dataset in the \code{data} ready to be exported}

\item{csv_in_pkg}{if \code{TRUE}, and when inside a package project, save the dataset as a csv file inside the \code{data-raw} directory}

\item{csv2zip}{if \code{TRUE} with also \code{csv_in_pkg} \code{TRUE}, and when inside a package project, zips the csv file inside the \code{data-raw} directory then deletes it}
}
\value{
none
}
\description{
Save a dataset in different formats, in various directories
}
\author{
Luca Valnegri, \email{l.valnegri@datamaps.co.uk}
}
