% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fst.R
\name{save_multi_idx}
\alias{save_multi_idx}
\title{Save a dataset multiple times with different indices}
\usage{
save_multi_idx(dts, fnames, fpath)
}
\arguments{
\item{dts}{the dataset to be saved}

\item{fnames}{a named list, where each name refers to the name of the file, while the content refers to the column(s) on which the index is built}

\item{fpath}{the directory where to store the files}
}
\value{
none
}
\description{
Save a dataset multiple times with different indices
}
\author{
Luca Valnegri, \email{l.valnegri@datamaps.co.uk}
}
