% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postcodes.R
\name{save_postcodes}
\alias{save_postcodes}
\title{Save the postcodes table in multiple files in fst format, each with its own index}
\usage{
save_postcodes(dts, pfn = FALSE)
}
\arguments{
\item{dts}{the dataset to be saved}

\item{pfn}{if TRUE, it saves also a version with index over PFA+PFN}
}
\value{
None
}
\description{
Save the postcodes table in multiple files in fst format, each with its own index
}
\author{
Luca Valnegri, \email{l.valnegri@datamaps.co.uk}
}
