% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{split_dts}
\alias{split_dts}
\title{Write a dataset in fst format with an index over one of its columns}
\usage{
split_dts(x, pct = 0.7, grouped = NULL)
}
\arguments{
\item{x}{the dataset to be split}

\item{pct}{percentage of the dataset to include as training}
}
\value{
a list with two data.tables
}
\description{
Write a dataset in fst format with an index over one of its columns
}
\author{
Luca Valnegri, \email{l.valnegri@datamaps.co.uk}
}
