% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sw.R
\name{star_wars_by_species}
\alias{star_wars_by_species}
\title{Star Wars by Species}
\usage{
star_wars_by_species(species)
}
\arguments{
\item{species}{character}
}
\value{
tibble
}
\description{
Return a tibble of Star Wars characters by species
}
\examples{
star_wars_by_species("Human")
star_wars_by_species("Droid")
star_wars_by_species("Wookiee")
star_wars_by_species("Rodian")
}
