% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{str_add_char}
\alias{str_add_char}
\title{str_add_char}
\usage{
str_add_char(x, n, chr = "0", pos = "l")
}
\arguments{
\item{x}{the string}

\item{n}{the desired \emph{total} number of characters to return}

\item{chr}{the character(s) to add}

\item{pos}{Indicate if \code{chr} should be inserted at the (l)eft (beginning of string) or (r)right (end of string)}
}
\value{
a string
}
\description{
Add a number of prefixed characters to one side of a string
}
\author{
Luca Valnegri, \email{l.valnegri@datamaps.co.uk}
}
