% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{str_dx}
\alias{str_dx}
\title{str_dx}
\usage{
str_dx(x, n)
}
\arguments{
\item{x}{the string}

\item{n}{the desired number of characters}
}
\value{
a string
}
\description{
Returns the last $n$ characters of a string
}
\author{
Luca Valnegri, \email{l.valnegri@datamaps.co.uk}
}
