% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{str_lf}
\alias{str_lf}
\title{str_lf}
\usage{
str_lf(x, n, pos = "r")
}
\arguments{
\item{x}{the string}

\item{n}{the desired number of characters}

\item{pos}{the side of the string to keep: \code{l}eft (the beginning) or \code{r}ight (the end)}
}
\value{
a string
}
\description{
Returns the first/last $n$ characters of a string
}
\author{
Luca Valnegri, \email{l.valnegri@datamaps.co.uk}
}
