% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps_leaflet.R
\docType{data}
\name{tiles.lst}
\alias{tiles.lst}
\title{tiles.lst}
\format{
An object of class \code{list} of length 23.
}
\usage{
tiles.lst
}
\description{
List of background tiles for `leaflet` maps
}
\references{
\url{https://leaflet-extras.github.io/leaflet-providers/preview/}. 

To work correctly, needs to be paired with `add_maptile`.
}
\keyword{datasets}
