% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{tmp_path}
\alias{tmp_path}
\title{Location for any temporary files}
\format{
An object of class \code{character} of length 1.
}
\usage{
tmp_path
}
\description{
Location for any temporary files
}
\keyword{datasets}
