% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{ukgeo.lst}
\alias{ukgeo.lst}
\title{List of UK geographic types}
\format{
An object of class \code{list} of length 1.
}
\usage{
ukgeo.lst
}
\description{
For the complete list see https://geoportal.statistics.gov.uk/search?collection=Dataset&sort=name&tags=all(PRD_RGC)
}
\keyword{datasets}
