% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{xval}
\alias{xval}
\title{Split a dataset for a $k$-fold Cross-Validation}
\usage{
xval(x, n)
}
\arguments{
\item{x}{the dataset to be split}

\item{n}{the number of folds to split}
}
\value{
a list with n items corresponding to the required $k$ folds
}
\description{
Split a dataset for a $k$-fold Cross-Validation
}
\author{
Luca Valnegri, \email{l.valnegri@datamaps.co.uk}
}
