% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geography.R
\name{add_geocodes}
\alias{add_geocodes}
\title{Check a character string is a UK postcode, then convert it to a 7-char format}
\usage{
add_geocodes(dt, clean_pc = TRUE, pc_cname = "postcode",
  oa_only = FALSE, census = TRUE, admin = TRUE, postal = TRUE,
  electoral = FALSE, nhs = FALSE, cols_in = NULL, cols_out = NULL)
}
\arguments{
\item{dt}{a data.table}

\item{cols_out}{The columns you don't want to be included in the output. You can not avoid to choose neither OA nor WPZ.}
}
\value{
a data.table with possibly cleaned postcode, OA and WPZ columns, plus all other specified
}
\description{
The UK postcode system is hierarchical, the top level being "Postcode Area" (PCA) identified by 1 or 2 alphabetical character.
The next level is the "Postcode District" (PCD), also commonly known as the "outcode", and can take on several different formats, and anywhere from 2 to 4 alphanumeric characters long.
Next comes the Postcode Sector" (PCS), always identified by a single number, then finally the "unit", always formed by two alphabetical characters.
The combination of "sector" and "unit" is often called "incode", which is always 1 numeric character followed by 2 alphabetical characters.
}
\examples{
\dontrun{
  add_geocodes(dts)
  add_geocodes(dts, postal = FALSE)
  add_geocodes(dts, postal = FALSE, workplace = TRUE)
}

}
\author{
Luca Valnegri, \email{l.valnegri@datamaps.co.uk}
}
