% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{add_pct}
\alias{add_pct}
\title{Format a decimal number as percentage, adding the % sign}
\usage{
add_pct(x, dd = 1)
}
\arguments{
\item{x}{a numeric vector}
}
\value{
A character vector
}
\description{
Format a decimal number as percentage, adding the % sign
}
\author{
Luca Valnegri, \email{l.valnegri@datamaps.co.uk}
}
