% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geography.R
\name{bounding_box}
\alias{bounding_box}
\title{Calculate a square or circle bounding box given a distance and a pair of coordinates}
\usage{
bounding_box(x_lon, y_lat, dist, in.miles = TRUE)
}
\arguments{
\item{x_lon}{The longitude of the center point.}

\item{y_lat}{The latitude of the center point.}

\item{dist}{The distance from the center point.}

\item{in.miles}{logical.  If \code{TRUE} uses miles as the units of \code{dist}. If \code{FALSE} uses kilometers.}
}
\value{
a matrix with two rows (the axis) and two columns (the limits)
}
\description{
Calculate a square or circle bounding box given a distance and a pair of coordinates
}
\references{
\url{http://janmatuschek.de/LatitudeLongitudeBoundingCoordinates}
}
\author{
Luca Valnegri, \email{l.valnegri@datamaps.co.uk}
}
