% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databases.R
\name{create_dbtable}
\alias{create_dbtable}
\title{Creates a MySQL table}
\usage{
create_dbtable(tname, dname, tdef, dts = NULL)
}
\arguments{
\item{tname}{the name of the table to be created}

\item{dname}{the name of the database in which create the table}

\item{tdef}{the definitions for all the columns in the table}

\item{dts}{the optional dataset to be uploaded}
}
\value{
None
}
\description{
\code{create_dbtable} creates a table in a specified database in the local MySQL Server
}
\author{
Luca Valnegri, \email{l.valnegri@datamaps.co.uk}
}
