% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geography.R
\name{dd_geocode}
\alias{dd_geocode}
\title{Geocode addresses in a data.table adding longitude and latitude coordinates using the Google Maps Engine}
\usage{
dd_geocode(
  y,
  na = "address",
  nx = "x_lon",
  ny = "y_lat",
  update_address = FALSE
)
}
\arguments{
\item{y}{The data.table with an \code{address} column}

\item{na}{The name for the \emph{longitude} column. If it already exists it will be deleted}

\item{nx}{The name for the \emph{longitude} column. If it already exists it will be deleted}

\item{ny}{The name for the \emph{latitude} column. If it already exists it will be deleted}

\item{update_address}{if \code{TRUE}, the address column will be updated with the result from the Google Maps Engine}
}
\value{
a data.table with two additional columns
}
\description{
Geocode addresses in a data.table adding longitude and latitude coordinates using the Google Maps Engine
}
\author{
Luca Valnegri, \email{l.valnegri@datamaps.co.uk}
}
