% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{get_map_legend}
\alias{get_map_legend}
\title{Build a legend to be used in a leaflet thematic map}
\usage{
get_map_legend(mtc, brks, dec.fig = 2, del_signs = TRUE)
}
\arguments{
\item{mtc}{the metric used in the choropleth}

\item{brks}{a vector that describes the $(n + 1)$ breaks that build the $n$ bins}

\item{dec.fig}{the number of decimal to keep}

\item{del_signs}{if TRUE, drop $+/-$ from numbers}
}
\value{
a character vector
}
\description{
Build a legend to be used in a leaflet thematic map
}
\author{
Luca Valnegri, \email{l.valnegri@datamaps.co.uk}
}
