% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{load_pkgs}
\alias{load_pkgs}
\title{load_pkgs}
\usage{
load_pkgs(..., uk = FALSE, it = FALSE, eu = FALSE, wd = FALSE)
}
\arguments{
\item{...}{other packages to be loaded}

\item{uk}{a logical to load the \code{UK} set of packages: 'RgeoUK', 'RbndUK', 'RdataUK'.}

\item{it}{a logical to load the \code{IT} set of packages: 'RgeoIT', 'RbndIT', 'RdataIT'.}

\item{eu}{a logical to load the \code{EU} set of packages: 'RgeoWD', 'RbndWD', 'RdataWD'.}

\item{wd}{a logical to load the \code{WD} set of packages: 'RgeoWD', 'RbndWD', 'RdataWD'.}
}
\value{
none
}
\description{
Load packages
}
\author{
Luca Valnegri, \email{l.valnegri@datamaps.co.uk}
}
