% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{load_pkgs}
\alias{load_pkgs}
\title{Load packages}
\usage{
load_pkgs(..., dmp = TRUE)
}
\arguments{
\item{dmp}{a logical vector for the dmpkg set: 'funs', 'geouk', 'bnduk', 'mapuk'.}

\item{y}{a character vector of packages to be loaded}
}
\value{
none
}
\description{
Load packages
}
\author{
Luca Valnegri, \email{l.valnegri@datamaps.co.uk}
}
