% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{marker_colours}
\alias{marker_colours}
\title{List of available fill colours for Awesome Markers in leaflet maps}
\format{
An object of class \code{character} of length 16.
}
\usage{
marker_colours
}
\description{
List of available fill colours for Awesome Markers in leaflet maps
}
\keyword{datasets}
