% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dates.R
\name{mnths}
\alias{mnths}
\title{Create a list of ordered months, possibly by year(s)}
\usage{
mnths(ys = NA, ye = NA)
}
\arguments{
\item{ys}{the start year, or the only year if \code{ye} is missing. If missing, the function returns only the months.}

\item{ye}{the final year}
}
\value{
a data.table
}
\description{
Create a list of ordered months, possibly by year(s)
}
\author{
Luca Valnegri, \email{l.valnegri@datamaps.co.uk}
}
