% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{palettes.lst}
\alias{palettes.lst}
\title{List of ColorBrewer palettes partitioned by type of visualization scale}
\format{An object of class \code{list} of length 3.}
\usage{
palettes.lst
}
\description{
List prepared to be used as choices in Shiny input controls
}
\keyword{datasets}
