% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{pub_path}
\alias{pub_path}
\title{Top root of the shared repository in every datamaps VPS}
\format{An object of class \code{character} of length 1.}
\usage{
pub_path
}
\description{
Top root of the shared repository in every datamaps VPS
}
\keyword{datasets}
