% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fst.R
\name{read_fst_idx}
\alias{read_fst_idx}
\title{Read a (partial) dataset from an fst indexed file based on values pertainig to one or two columns}
\usage{
read_fst_idx(fname, ref, cols = NULL)
}
\arguments{
\item{fname}{the name of the \code{fst} file complete with its path}

\item{ref}{the value(s) pertaining to the column(s) that makes up the index}

\item{cols}{the columns to be returned (the \code{NULL} default means to return all columns)}
}
\value{
A data.table
}
\description{
Read a (partial) dataset from an fst indexed file based on values pertainig to one or two columns
}
\author{
Luca Valnegri, \email{l.valnegri@datamaps.co.uk}
}
