% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rfuns.R
\docType{package}
\name{Rfuns}
\alias{Rfuns}
\title{Rfuns: a bunch of functions to help Data Sourcing, Cleaning and Wrangling for data projects}
\description{
The `Rfuns` package provides the following categories of important functions:
 - connection and querying to local MySQL and main SQL Server databases
 - cleaning, recoding, saving, and querying R datasets
 - quick summaries and mapping
}
