% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boundaries.R
\name{save_bnd}
\alias{save_bnd}
\title{save_bnd}
\usage{
save_bnd(bnd, bname, bpath, pct = "00", shp = TRUE, rds = TRUE)
}
\arguments{
\item{bnd}{the \code{sf} object to be saved}

\item{bname}{the name of the output file}

\item{bpath}{the \emph{root} output folder where to store the file(s)}

\item{pct}{a 2-char string identifying the percentage of simplification: must be '05', '10', '20', '30', '40', '50'. '00' stand for "original size"}

\item{shp}{if true, save as shapefile}

\item{rds}{if true, save in RDS format}

\item{qs}{if true, save as shapefile}
}
\value{
none
}
\description{
Save an \code{sf} object as a \emph{shapefile} and/or in \code{qs} format and/or in \code{RDS} format
}
\author{
Luca Valnegri, \email{l.valnegri@datamaps.co.uk}
}
