% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{save_dts_pkg}
\alias{save_dts_pkg}
\title{save_dts_pkg}
\usage{
save_dts_pkg(
  y,
  fn,
  fst_path,
  as_rdb = TRUE,
  dbn = NULL,
  tbn = fn,
  as_rda = TRUE,
  csv_in_pkg = TRUE,
  csv2zip = FALSE
)
}
\arguments{
\item{y}{a dataset}

\item{fn}{a common name for all the file formats}

\item{fst_path}{the path where to save the file in `fst` format (usually inside the *public* repository)}

\item{as_rdb}{if `TRUE`, save the dataset in the `dbn` database.}

\item{dbn}{the name of the database to be used when `as_rdb` is `TRUE`}

\item{tbn}{The name for the database table where to store the dataset (by default equals the filename).
The table should be conveniently created beforehand.
The current table will be truncated, so that existing data will be deleted}

\item{as_rda}{if `TRUE`, and when inside a package project, save the dataset in the `data` ready to be exported}

\item{csv_in_pkg}{if `TRUE`, and when inside a package project, save the dataset as a csv file inside the `data-raw` directory}

\item{csv2zip}{if `TRUE` with also `csv_in_pkg` `TRUE`, and when inside a package project, zips the csv file inside the `data-raw` directory then deletes it}
}
\value{
none
}
\description{
Save a dataset in different formats in various directory
}
\author{
Luca Valnegri, \email{l.valnegri@datamaps.co.uk}
}
