% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{tiles.lst}
\alias{tiles.lst}
\title{List of background tiles for leaflet maps}
\format{An object of class \code{list} of length 15.}
\usage{
tiles.lst
}
\description{
For the complete list see https://leaflet-extras.github.io/leaflet-providers/preview/
}
\keyword{datasets}
