% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fst.R
\name{write_fst_idx}
\alias{write_fst_idx}
\title{Write a dataset in fst format with an index over one of its columns}
\usage{
write_fst_idx(tname, cname, dts = NA, out_path = "./", fname = NA, dname = NA)
}
\arguments{
\item{dname}{}
}
\value{
None
}
\description{
Write a dataset in fst format with an index over one of its columns
}
\author{
Luca Valnegri, \email{l.valnegri@datamaps.co.uk}
}
